Button to delete class
