+++ typescript/web/src/components/labelling-tool/options-tool-bar/options-tool-bar.tsx
@@ -1,3 +1,9 @@
+import { DeleteLabelButton } from "./delete-label-button";
 import { EditLabelClassMenu } from "./edit-label-class-menu";
 
-export const OptionsToolBar = () => <EditLabelClassMenu />;
+export const OptionsToolBar = () => (
+  <>
+    <EditLabelClassMenu />
+    <DeleteLabelButton />
+  </>
+);
