+++ typescript/web/src/components/labelling-tool/zoom-tool-bar/zoom-tool-bar.tsx
@@ -1,6 +1,7 @@
 import {
   IconButton,
   chakra,
+  Tooltip,
   useColorModeValue as mode,
 } from "@chakra-ui/react";
 import { RiZoomInLine, RiZoomOutLine } from "react-icons/ri";
@@ -18,26 +19,30 @@ export const ZoomToolbar = () => {
 
   return (
     <>
-      <IconButton
-        icon={<ZoomInIcon fontSize="lg" />}
-        backgroundColor={mode("white", "gray.800")}
-        aria-label="Zoom in"
-        pointerEvents="initial"
-        isDisabled={!canZoomIn}
-        onClick={() => {
-          zoomByDelta(zoomFactor);
-        }}
-      />
-      <IconButton
-        icon={<ZoomOutIcon fontSize="lg" />}
-        backgroundColor={mode("white", "gray.800")}
-        aria-label="Zoom out"
-        pointerEvents="initial"
-        isDisabled={!canZoomOut}
-        onClick={() => {
-          zoomByDelta(-zoomFactor);
-        }}
-      />
+      <Tooltip label="Zoom In" placement="left" openDelay={300}>
+        <IconButton
+          icon={<ZoomInIcon fontSize="lg" />}
+          backgroundColor={mode("white", "gray.800")}
+          aria-label="Zoom in"
+          pointerEvents="initial"
+          isDisabled={!canZoomIn}
+          onClick={() => {
+            zoomByDelta(zoomFactor);
+          }}
+        />
+      </Tooltip>
+      <Tooltip label="Zoom Out" placement="left" openDelay={300}>
+        <IconButton
+          icon={<ZoomOutIcon fontSize="lg" />}
+          backgroundColor={mode("white", "gray.800")}
+          aria-label="Zoom out"
+          pointerEvents="initial"
+          isDisabled={!canZoomOut}
+          onClick={() => {
+            zoomByDelta(-zoomFactor);
+          }}
+        />
+      </Tooltip>
     </>
   );
 };
